Transfer method (createrawtx, decoderawtx, getrawtx, sendrawtx) to transacion space
