+++ commands/box/transaction/root.go
@@ -5,6 +5,7 @@
 package transactioncmd
 
 import (
+	"encoding/hex"
 	"fmt"
 	"path"
 	"strconv"
@@ -14,6 +15,7 @@ import (
 	"github.com/BOXFoundation/boxd/config"
 	"github.com/BOXFoundation/boxd/core"
 	"github.com/BOXFoundation/boxd/core/types"
+	"github.com/BOXFoundation/boxd/crypto"
 	"github.com/BOXFoundation/boxd/rpc/rpcutil"
 	"github.com/BOXFoundation/boxd/util"
 	"github.com/BOXFoundation/boxd/wallet"
@@ -64,7 +66,180 @@ func init() {
 				fmt.Println("signrawtx called")
 			},
 		},
+		&cobra.Command{
+			Use:   "createrawtx [from] [txhash1,txhash2...] [vout1,vout2...] [to1,to2...] [amount1,amount2...]",
+			Short: "Create a raw transaction",
+			Run:   createRawTransaction,
+		},
+		&cobra.Command{
+			Use:   "decoderawtx",
+			Short: "A brief description of your command",
+			Long: `A longer description that spans multiple lines and likely contains examples
+and usage of using your command. For example:
+
+Cobra is a CLI library for Go that empowers applications.
+This application is a tool to generate the needed files
+to quickly create a Cobra  application.`,
+			Run: decoderawtx,
+		},
+		&cobra.Command{
+			Use:   "getrawtx [txhash]",
+			Short: "Get the raw transaction for a txid",
+			Run:   getRawTxCmdFunc,
+		},
+		&cobra.Command{
+			Use:   "sendrawtx [rawtx]",
+			Short: "Send a raw transaction to the network",
+			Run:   sendrawtx,
+		},
 	)
+}
+
+func createRawTransaction(cmd *cobra.Command, args []string) {
+	if len(args) < 4 {
+		fmt.Println("Invalide argument number")
+		return
+	}
+	fmt.Println("createRawTx called")
+	from := args[0]
+	//Cut characters around commas
+	txHashStr := strings.Split(args[1], ",")
+	txHash := make([]crypto.HashType, 0)
+	for _, x := range txHashStr {
+		tmp := crypto.HashType{}
+		if err := tmp.SetString(x); err != nil {
+			fmt.Println("set string error: ", err)
+			return
+		}
+		txHash = append(txHash, tmp)
+	}
+	voutStr := strings.Split(args[2], ",")
+	vout := make([]uint32, 0)
+	for _, x := range voutStr {
+		tmp, err := strconv.Atoi(x)
+		if err != nil {
+			fmt.Println("Type conversion failed: ", err)
+			return
+		}
+		vout = append(vout, uint32(tmp))
+	}
+	toStr := strings.Split(args[3], ",")
+	to := make([]string, 0)
+	for _, x := range toStr {
+		to = append(to, x)
+	}
+	amountStr := strings.Split(args[4], ",")
+	amounts := make([]uint64, 0)
+	for _, x := range amountStr {
+		tmp, err := strconv.Atoi(x)
+		if err != nil {
+			fmt.Println("Type conversion failed: ", err)
+			return
+		}
+		amounts = append(amounts, uint64(tmp))
+	}
+	if len(txHash) != len(vout) {
+		fmt.Println(" The number of [txid] should be the same as the number of [vout]")
+		return
+	}
+	if len(to) != len(amounts) {
+		fmt.Println("The number of [to] should be the same as the number of [amount]")
+		return
+	}
+	conn, err := rpcutil.GetGRPCConn(getRPCAddr())
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	defer conn.Close()
+	height, err := rpcutil.GetBlockCount(conn)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	tx, err := rpcutil.CreateRawTransaction(from, txHash, vout, to, amounts, height)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	mashalTx, err := tx.Marshal()
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	strTx := hex.EncodeToString(mashalTx)
+	fmt.Println(strTx)
+}
+
+func sendrawtx(cmd *cobra.Command, args []string) {
+	fmt.Println("sendrawtx called")
+	if len(args) != 1 {
+		fmt.Println("Can only enter one string for a transaction")
+		return
+	}
+	conn, err := rpcutil.GetGRPCConn(getRPCAddr())
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	defer conn.Close()
+	txByte, err := hex.DecodeString(args[0])
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	tx := new(types.Transaction)
+	if err := tx.Unmarshal(txByte); err != nil {
+		fmt.Println("unmarshal error: ", err)
+		return
+	}
+	resp, err := rpcutil.SendTransaction(conn, tx)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	fmt.Println(util.PrettyPrint(resp))
+}
+
+func getRawTxCmdFunc(cmd *cobra.Command, args []string) {
+	fmt.Println("getrawtx called")
+	if len(args) < 1 {
+		fmt.Println("Param txhash required")
+		return
+	}
+	hash := crypto.HashType{}
+	hash.SetString(args[0])
+	conn, err := rpcutil.GetGRPCConn(getRPCAddr())
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	defer conn.Close()
+	tx, err := rpcutil.GetRawTransaction(conn, hash.GetBytes())
+	if err != nil {
+		fmt.Println(err)
+	} else {
+		fmt.Println(util.PrettyPrint(tx))
+	}
+}
+
+func decoderawtx(cmd *cobra.Command, args []string) {
+	fmt.Println("decoderawtx called")
+	if len(args) != 1 {
+		fmt.Println("Invalide argument number")
+		return
+	}
+	txByte, err := hex.DecodeString(args[0])
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	tx := new(types.Transaction)
+	if err := tx.Unmarshal(txByte); err != nil {
+		fmt.Println("Unmarshal error: ", err)
+		return
+	}
+	fmt.Println(util.PrettyPrint(tx))
 }
 
 func sendFromCmdFunc(cmd *cobra.Command, args []string) {
